Make it work in DF
In order for EasyDestory to work in Dragonflight a number of api changes
needed to be made to interface with the container item methods that were
namespaced (and updated).

Additionally a number of ENUMS were updated at some point which may have
been the source of people not seeing items. Seems strange that the issue
didn't impact me until DF though others apparently were suffering from
the issue earlier. Maybe something on Blizzards end?